src/Customer Service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
alembic==1.7.1
email-validator
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-zipkin
dapr-ext-fastapi>=1.0.0
dapr>=1.0.0 
